Add format attribute to drv_log_msg
Prior to this change, clang was throwing a warning on the call to
vfprintf, stating that the `fmt` argument was not a string literal. But
this is supposed to be the case, since fmt is passed in from the caller
of this function.

This change adds the `format` attribute to drv_log_msg. This is
supported in both GCC and Clang, and will prevent either compiler from
erroring out on the call to vfprintf.